// src/sdks/Newcomputation/computationV2ApiDocs.ts
export type getComputationReportPageProps = {
  likeReportName?: string | null | undefined;
  orgId?: number | null | undefined;
  pageNum: number;
  pageSize: number;
  year?: number | null | undefined;
};
// ...
export interface Report {
  absolution?: string;

  biomassDesc?: string;

  carbonEmission?: number;

  companyId?: number;

  controlPlanId?: number;

  controlPlanVersion?: string;

  createBy?: number;

  createTime?: Date;

  deleted?: boolean;

  elseDesc?: string;

  id?: number;

  mainEmail?: string;

  mainName?: string;

  mainPhone?: string;

  orgId?: number;

  orgName?: string;

  quantizationChangeDesc?: string;

  reductionSceneIds?: number[];

  reportName?: string;

  standardEmission?: number;

  updateBy?: number;

  updateByName?: string;

  updateTime?: Date;

  year?: number;
}